feat: log total records, not only primary data
When we logged at the controller level, we only got a sense of how much
primary data was included in a query. However, the included data can be as
large or larger and we didn't track it.

By pulling the logging into the ApiViewHelpers, we can do the log after we've
processed the included data as well.