feat(server): hibernate after messages
For most of the Server instances, they only receive messages at startup and
then are idle. By hibernating, we allow the memory they used to parse the
GTFS data to be garbage collected.

For predictions/vehicles/alerts, they receive many messages to reset their
state. We don't want to hibernate every time for those, so we create a new
option for the server which allows us to configure whether we hibernate after messages.