chore: update tzdata/fast_local_datetime/timex
There's a bug in earlier versionf of tzdata when processing the 2019b
timezone database (luckily not in code I wrote :). This updates us to the
latest version of our time dependencies, and fixes the issues that came up.

The bug one looks to be an issue with Timex and how it handle DST
transitions. In our previous test, we checked that our parsing returned the
same results as Timex. Now Timex is returning the wrong data, so we switch to
a simpler test that we're returning a DateTime that has the same epoch timestamp.