feat: add stop query for location_type 0 stops
When we're querying for schedules or predictions, the vehicles can only go to
stops with a location_type of 0. However, we weren't filtering out other
types of stops when doing those queries, making the list of stops to search
for very large.

This new helper function filters the resulting list of IDs to only those with
location_type 0, making it easier for Schedules and Predictions in the future
to query for only the relevant stop IDs.