feat(Prediction): store/query route_pattern_id
Since the website is using this query very frequently, it benefits the
queries dramatically to be able to use it as a direct index, rather than
looking it up from the Trips. It'll also allow us to more easily use the
route_pattern_id provided by upstream services in the future.