perf: ScheduleView rendering
JaSerializer already filters the fields later on, so it's duplicated effort
for us to use `Map.take` to filter them out. We still want to to avoid
calcuating the times if we can avoid it.

If we ensure that `opts` is always a map, then we can do easier pattern
matching as well.