fix: create an Added trip whenver we can't match the trip ID
Previously, we'd only create an Added trip if the schedule relationship was
`nil`, meaning scheduled. However, in some cases where the prediction data
doesn't line up with the current GTFS data, we can have scheduled trips but
no trip data from GTFS. In that case, we still want to provide a trip if
possible, so we treat it similarly to an ADDED trip, and generate a record
with the last predicted stop as the headsign.