test: Fix broken test exposed by `plug` update
Some versions of `plug` prior to v1.10.2 had a bug where `query_string`
was not correctly populated in tests. That caused this test to pass when
it should have failed, because the passed `date` filter never made it to
the controller. With the bug fixed, the test crashes in `ValidateDate`,
though since the crash is because the feed isn't loaded (and is never
loaded by default in tests), it's unclear why this wouldn't have caused
an issue before.

In any case, since the date filter doesn't appear to be meaningful here,
it can be removed and the test data can use a date that falls within the
test Service.