fix: remove unused `date` filter from predictions
The `/predictions` endpoint allowed a `filter[date]` parameter, but
didn't do anything with it. This is a backwards-incompatible change,
since previously-working requests that included this parameter will now
fail with a "no such filter" error.

This also clarifies the ambiguously-named `date` plug in the same
controller, and makes it private since no code outside the controller
called it.