test: more flexible Kingston/Plymouth validation
To guard against regression of a bug which caused Plymouth to not be
present in the `CR-Kingston` stop list on weekends, this test asserted
that it should be there on every day of service. However, this prevents
us from implementing a disruption where the Kingston line actually will
not stop at Plymouth for a few days, due to the station being closed.

This changes the test to check the stop list by service (instead of by
date, which was translated to a service internally), and to only check
services with a `schedule_typicality` less than 4. This allows services
that don't stop at Plymouth (or Kingston) to pass validation as long as
they're considered "atypical".