chore: handle load balancer connection draining
When an API instance is shut down on AWS, the following sequence of
events occurs:

1. ALB stops routing incoming connections to the instance ("draining")
2. ALB waits for a configurable amount of time
3. ALB forcibly closes any remaining open connections to the instance
4. ECS sends SIGTERM to the application

This creates a problem where the app doesn't know it's being shut down
until it's already too late to cleanly close any open connections. This
isn't an issue for most apps since HTTP connections generally have a
lifetime shorter than the ALB draining time, but it is an issue for
event-stream connections, which remain open indefinitely (until either
the client closes them or an error occurs). When ALB closes these, it
creates unwanted errors for clients.

To work around this, we add an `ALBMonitor` process to the app, which
periodically checks the instance's status in the ALB API. If the status
is "draining", we proactively shut down the VM so the app can clean up
its own connections before ALB does. Phoenix performs its own connection
draining on shutdown, and in the commit prior to this we added "shutdown
awareness" to event streams.