fix: close event streams on app shutdown
We handle event streams using a Plug that (under normal circumstances)
never returns — the Phoenix process spawned to handle the connection
would thus remain alive until terminated as part of app shutdown. The
HTTP connection would not be closed cleanly in this case, resulting in
clients getting errors.

To address this, we add a "Canary" process after the Endpoint in the
supervision tree, whose sole purpose is to be terminated before the
Endpoint so it can give the EventStream processes advance notice of the
shutdown. Currently it does so by terminating all running DiffServers,
which EventStreams now detect and respond by closing their connections.

Event streams also now send an error response and close the connection
if their DiffServer crashes for any reason; previously they would stay
alive but stop delivering new data.