refactor: remove double-negative in stops_on_route
Certain trips can be "ignored" by the default logic that determines
which stops are "on" a route. We created a config value to override
this, but due to "ignored" being `true` internally, a `false` in this
config meant "don't ignore", or "don't _not_ consider these trips" — a
double negative that made the value confusing to read.

This inverts the meaning of the config, such that `true` means "_do_
consider these trips", and propagates that change through the rest of
the code:

* "canonical" is used as the opposite of "ignored" (this also replaces
  "alternate" in `StopsOnRoute` specifically, which referred to the fact
  that originally alternate-route trips were the only "ignored" ones)

* `ignore_trip_route_pattern?` and `ignore_trip_for_route?` are renamed
  `stops_on_route?` and `stops_on_route_by_shape?`, to clarify they have
  the same basic function except for the latter using the shape priority
  (and this is the unusual one, since shape priorities are deprecated)