fix: Remove unused/broken code
Elixir 1.12 was able to realize that `input_validations` had a bug in it.
The function header guarded for %ApiAccounts.Changeset{} structs. However,
the code then did changeset.validations and changeset.errors, which are not
fields on this struct.

The code was probably expecting to be an Ecto changeset, which does have
those fields, and indeed in the commit where it was added, the function
head used an unnamed map. However, the ApiAccounts module was still "TODO"
in that commit. I suspect plans changed and this didn't quite work out as
expected.

This code has never been functional, but the protocol requires the function
to be defined, and will raise a warning if it's not. So I've replaced the code
with something minimal.