fix: better enforce restrictions on `occupancies`
The `occupancies` relationship on Trips was only intended to be
accessible when the client has opted into experimental features, but
this restriction only worked when trips were accessed via the `/trips`
endpoints. If loading trips via another resource's endpoints, such as
`/schedules?include=trip.occupancies`, the restriction did not apply.

This patch addresses the issue in basically the same way we handle
other "conditionally existing" relationships: if accessed indirectly,
the request will still be allowed, but the relationship will not
contain any data. This does not attempt to address the broader
inconsistency that validation of the `include` parameter only works at
the "top level" of the include tree.