fix: gracefully shut down connections with a plug
This introduces a new plug, `CheckForShutdown`. With each request, we
check a `:persistent_term` variable for whether the application is
shutting down. If we are, then we set the `connection: close` header,
which tells the ALB and other HTTP clients to terminate the connection.

`:persistent_term` (introduced with Erlang 21) is a fast way of setting
global variables. Some local benchmarking indicates that it takes
approximately 0.01us (microsecond) for the call, so it's not adding
overhead to do it for each request.

We also update the Canary (now `ApiWeb.Canary`) to shutdown the plug as
well as the EventStream clients.