fix: also gracefully shut down EventStream
While we do terminate any running EventStream instances at shutdown,
it's possible for some new ones to sneak in before the listener is
suspended. To handle that case, we check whether the shutdown has
happened after processing any events, and close the connection if that
happens. Even in the case where there's no updates to the feed, the
`:timeout` message should happen at least every 30s, giving the
connection a change to disconnect.