perf: special case searching with a single Stop filter
The existing behavior for `State.Stop.filter_by/1` when there is at
least one filter is to run the filters one-at-a-time, filtering out any
stops which were already present in the list.

If there's only a single filter, we don't need to do any of that extra
work: we can simply return the value from the single filter.