fix(vehicles): add logging if a non-shuttle vehicle has a null direction_id (#854)
Problem: As described in the linked ticket, sometimes we return vehicles with "direction_id": null. We aren't sure why. I spent a couple of days looking into this issue and didn't make any progress.

Solution: We are adding logs for when an invalid direction_id is returned. Hopefully these can allow us to better ascertain what conditions lead to this happening.